requirements.txt
yfinance
pandas
numpy
openpyxl
gspread
gspread-dataframe
google-auth
